Remove unnecessary casts for NumPy limits
Here we change those conversions of the numpy.finfo structure
where the type does not change (e.g. float64 values). This is
possibly only correct for our target OS platforms, but given
the behavior of Python, and it's type conversions, it makes
sense to avoid complicating the code.

+++ sherpa/fit.py
@@ -724,7 +724,7 @@ class IterFit(NoNewAttributesAfterInit):
         # Keep record of current and previous statistics;
         # when these are within some tolerace, Primini's method
         # is done.
-        previous_stat = np.float32(np.finfo(np.float32).max)
+        previous_stat = np.finfo(np.float32).max
         current_stat = statfunc(pars)[0]
         nfev = 0
         iters = 0
